+++ sdk/base-sys/src/bindings/active_account.rs
@@ -1,6 +1,6 @@
 use miden_stdlib_sys::{Felt, Word, WordAligned};
 
-use super::types::{AccountId, RawAccountId};
+use super::types::{AccountId, Nonce, RawAccountId};
 
 #[allow(improper_ctypes)]
 unsafe extern "C" {
@@ -66,8 +66,10 @@ pub fn get_id() -> AccountId {
 
 /// Returns the nonce of the active account.
 #[inline]
-pub fn get_nonce() -> Felt {
-    unsafe { extern_active_account_get_nonce() }
+pub fn get_nonce() -> Nonce {
+    Nonce {
+        inner: unsafe { extern_active_account_get_nonce() },
+    }
 }
 
 /// Computes and returns the commitment of the current account data.
@@ -193,7 +195,7 @@ pub trait ActiveAccount {
 
     /// Returns the nonce of the active account.
     #[inline]
-    fn get_nonce(&self) -> Felt {
+    fn get_nonce(&self) -> Nonce {
         self.__assert_active_account();
         get_nonce()
     }