+++ sdk/base-sys/src/bindings/native_account.rs
@@ -1,6 +1,6 @@
 use miden_stdlib_sys::{Felt, Word, WordAligned};
 
-use super::types::{AccountId, Asset, RawAccountId};
+use super::types::{AccountId, Asset, Nonce, RawAccountId};
 
 #[allow(improper_ctypes)]
 unsafe extern "C" {
@@ -150,8 +150,10 @@ pub fn get_id() -> AccountId {
 
 /// Increments the account nonce by one and returns the new nonce.
 #[inline]
-pub fn incr_nonce() -> Felt {
-    unsafe { extern_native_account_incr_nonce() }
+pub fn incr_nonce() -> Nonce {
+    Nonce {
+        inner: unsafe { extern_native_account_incr_nonce() },
+    }
 }
 
 /// Computes and returns the commitment to the native account's delta for this transaction.
@@ -277,7 +279,7 @@ pub trait NativeAccount {
 
     /// Increments the account nonce by one and returns the new nonce.
     #[inline]
-    fn incr_nonce(&mut self) -> Felt {
+    fn incr_nonce(&mut self) -> Nonce {
         incr_nonce()
     }